requirements.txt
asgiref==3.11.0
Django==6.0.1
django-taggit==5.0.1
Markdown==3.6
psycopg==3.1.18
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0
